Optimize statistics check via switchpoint
References: https://github.com/scala/scala-dev/issues/149.

It adds a mechanism to remove the overhead of checking for statistics in
the common case (where statistics are disabled). It does so by reusing a
mechanism proposed by Rémi Forax and brought to my attention by
Jason (more here: https://github.com/scala/scala-dev/issues/149).

This commit is work-in-progress and I've just opened it to check the
performance penalty of my changes. Afterwards, I would like to detect
what's the performance penalty of having statistics enabled by default,
I'm sure we can draw interesting results from there.

There is one pending issues with this commit, which hasn't been solved yet:

* How do we keep this implementation runtime-agnostic, allowing
  compilers like Scalajs and Scala.native to default to an inefficient
  implementation?

If you like the implementation, I could look into ways of expanding this
implementation to all our flag checking infrastructure, which is what
the referenced ticket on the top is about. However, I'd like to do this
in a separate PR if possible. For now, I wanted to create a very small
change to see the prior results.